test: modularize common test class
Create a TSS2_BaseTest class that you can import, and just run your
tests on.
